rename to git-utils.js #523
