+++ src/deploy.cmd.js
@@ -21,7 +21,7 @@ const fs = require('fs-extra');
 const uuidv4 = require('uuid/v4');
 const ProgressBar = require('progress');
 const { HelixConfig, GitUrl } = require('@adobe/helix-shared');
-const GitUtils = require('./GitUtils');
+const GitUtils = require('./git-utils');
 const useragent = require('./user-agent-util');
 const AbstractCommand = require('./abstract.cmd.js');
 const PackageCommand = require('./package.cmd.js');
