+++ src/publish.cmd.js
@@ -19,7 +19,7 @@ const path = require('path');
 const glob = require('glob-to-regexp');
 const { toBase64 } = require('request/lib/helpers');
 const ProgressBar = require('progress');
-const GitUtils = require('./GitUtils');
+const GitUtils = require('./git-utils');
 const strainconfig = require('./strain-config-utils');
 const include = require('./include-util');
 const useragent = require('./user-agent-util');
