+++ test/testDeployCli.js
@@ -16,7 +16,7 @@
 
 const assert = require('assert');
 const sinon = require('sinon');
-const GitUtils = require('../src/GitUtils');
+const GitUtils = require('../src/git-utils');
 const CLI = require('../src/cli.js');
 const DeployCommand = require('../src/deploy.cmd');
